Add new endpoint and update assembly version
Added a new GET endpoint `FetchIdsFromUsername` in `UserAuthenticationController.cs` to fetch user and profile IDs based on the username. This endpoint is protected by the `Authorize` attribute with the "Bearer" authentication scheme. Updated `AssemblyInfo.cs` to change the `AssemblyInformationalVersionAttribute` to a new version. Updated `AssemblyInfoInputs.cache` to reflect a new cache key.